Use fakeroot with make-kpkg
In the past, we've had weird issues with compiling kernels multiple
times on the same machine. Often, the first kernel build succeeds, but
then subsequent builds fail. The failure is strange. There are no
apparent error messages, and the build exits with a success exit
code. However, the build typically exits in ~20 minutes (as opposed to
the expected 2hr+ build time) and the `linux-image-*.deb` package is
much smaller than it should be (usually ~6MB instead of the expected
40MB).

A lot of kernel build documentation recommends avoiding building kernels
as root, which we had been doing, for reasons unclear. They instead
recommend using commands like `fakeroot` when building the kernel.

I tested our build process with `fakeroot` and was able to perform
multiple successful builds on the same VM. This suggests that fakeroot
may resolve the weird build issues; however, those issues were
intermittent, and time-consuming to reproduce. We should keep an eye out
for future occurrences of the same (or similar) issues because this
commit might not be a complete fix.