Uses OS detection to identify package list.
Supports Debian and Ubuntu, with slightly different
apt lists for both. Really, the apt list for Ubuntu
is specific to 14.04 LTS, so maybe it'd be wise to do
additional detection.

The role will fail if compilation is attempted on a
distro other than Debian or Ubuntu, with a helpful
message describing the problem and soliciting a PR. =)