Changes download directory location for building.
The previous value was `/usr/src`, of which the Linux FHS says
"Don't work here with your own projects". There's a separate
entry for `/usr/src/linux`, but that's no good either:

  You should probably use another directory
  when building your own kernel.

About `/usr/local/share`, the guidance is simply
"Source code for locally installed software",
which sounds downright friendly in the context
of the FHS, so let's use that.