Dynamic variable for kernel deb package.
Rather than hardcode the kernel version number
in the filepath, use the version variable
to structure a var for the filename, to make fetching
more reliable.

The value is still duplicated in a few places, which I don't
like, but the top-level `vars.yml` provides a clean way
to update the values going forward.