Renames var package_name -> grsecurity_deb_package
Requires hardcoding the filepath in the install playbooks.
It's not possible to dynamically declare the var in the
top-level vars file, because the components such as
`linux_kernel_version` are determined conditionally
within the build role.

It may be worthwhile to separate the variable resolution
logic into a distinct role, and set that role as a dependency
for the other roles, so they're all using the same vars.
I'm not going to do that now, so hardcoding the deb packagei
is the safest way to go.

Besides, the goal is to have these packages installable via
an apt repo anyway.