Uses unarchive module where possible.
Still need a `command` call to `xz`, since the
the `unarchive` module overhelpfully removes
intermediate archive stages. For this role, we need
the .tar file around to verify the GPG sig,
so we do a two-pass extraction.

Also converted `unxz` -> `xz --decompress`, because the
xz manpage says:

    When writing scripts that need to decompress files, it is
    recommended to always use the name xz with appropriate
    arguments (xz -d or xz -dc) instead of the names unxz and
    xzcat.

Maddeningly, it says that immediately after it says this:

    unxz is equivalent to xz --decompress.

Let no one say I did not read the manpage.