Beefs up Vagrant VM for building kernel.
Defaults to using all cores on host, but provides
a convenient way to halve that number, in case you want
to work while you build. Hardcodes 2GB of RAM for the guest VM.

Also switches to using the ubuntu/trusty64 boxes, which
reduce the number of packages that must be pulled down
via an apt-get upgrade.

Adds a secondary VM for testing installation, off by default.