Defaults to "test" patches in Vagrantfile
In the spirit of maximizing usefulness for first-timers trying out the
grsecurity patches, let's default to using the "test" patches, since
that playbook does not prompt for grsecurity auth credentials, which are
required for the stable patches.

In order to use Vagrant to build the stable patches, a simple one-line
change is necessary to swap the playbook order. Conceivably we could
support environment variables to toggle playbooks, but it seems more
reasonable to ship examples and expect folks to customize those in their
own Ansible project directories, which custom inventories.