Removes paxtest package from install requirements
We're not actually using paxtest anywhere in these roles. Down the road,
it may be worthwhile to add it back and use it to validate that the
grsecurity-patched kernel provides the protections we're looking for,
via the `paxtest blackhat` command. Tests (forthcoming) should use the
same method to validate the installation role.