Makes paxctl flag manipulation optional (on by default)
Internally we're experimenting with a role for paxctld that replaces the
need for using the paxctl utility. In fact, the two should NOT be used
together, since they use different methods for setting the headers.

Keep the sane default of setting the flags on the grub binaries, but
permit disabling the paxctl usage, in case paxctld is preferred.