Fails out of build role if running with superuser privileges
The Linux kernel dev community is unequivocal about the fact that while
installing the kernel requires superuser privileges, building the kernel
absolutely does not, and ignoring this fact can break your system.

Failing out of the role is a bit aggressive, but what minor irritation
that might cause is surely far less than making the build host
unbootable. Better safe than sorry.

Additionally, there's the stylistic matter that most Ansible users
assume that roles should be run with "become: true" either on the play
or on the role include. Perhaps it would be saner to support both, and
set "become: no" on the relevant compile tasks.