Smarter inference of privileges on remote host
The task previously referenced "ansible_ssh_user" and "ansible_user_id"
as a surrogates for checking whether the role was run with "sudo". As of
Ansible v2.0.2, however, "ansible_ssh_user" is now "null" unless
explicitly set, so we can't assume it'll be defined any longer. Let's
fall back to "ansible_env.USER", which requires fact gathering, but is
more resilient.

The slight modification to the folding scalar operator in the "msg"
parameter of the debug task is to strip the trailing newline on the
message in Ansible v2. The change is backwards compatible with v1.