Makes fetching back built packages optional
The primary intended use case for this role is running a build on a
remote server, in which case the automatic fetching back of .deb
packages makes sense. The role is equally useful for compiling kernels
on a local workstation, however, and in that case fetching the packages
back doesn't make sense.