Supports installation of multiple deb packages
The grsec-install role previously only supported installing a single
.deb package, intended for the kernel-image package. If the
kernel-headers, kernel-src, or doc or debug packages should also be
installed, a list var is necessary. Edited the install logic to retain
the install-only-if-recently-copied behavior.

Support for the original singular string for deb package remains in
place; the install task will include that string, if populated, as well
as anything in the list.