Ensure configure tasklisk skipped when no custom config is passed
Setting grsecurity_build_custom_config to an empty str instead of None matches
the conditional statments in the main.yml of the grsec-build role. This way, the
configure and compile taskslists get skipped, and the running the role with the
manual strategy works as intended, instead of ending in failure.