Creates kernel postinst hook in metapackage
We must ensure that the PaX flags are set correctly on the grub
binaries, otherwise the `zz-update-grub` postinst kernel hook will fail
on servers where both of the following are true:

  1. already running a grsec kernel
  2. have had grub updated since last SD Ansible playbook run

The longterm fix should include a solution to manage PaX flags over
time, e.g. paxctld or paxrat, but we still need a way to reliably update
running systems and ensure the newly installed kernel-image package
becomes the default.