Bumps metapackage version number with suffix
Using the "+r1" suffix appended to the version for the
`securedrop-grsec` metapackage because otherwise we won't have a unique
Version field, meaning automatic updates won't install the new package.
The previous Version field was simply "3.14.79", the same as the Linux
source version, and since we're reusing that source (but with a newer
grsecurity patch), we need a unique identifier.