Creates default build role var for revision
Using a default var so it's both overrideable, and so that other roles in
the same play can reuse the var (e.g. metapackage role). By default,
`make-kpkg` will use the revision "10.00.Custom", which can introduce
breakage in the update flow if you compile with a different grsecurity
patch set but the same kernel source (as sometimes happens).

Therefore in the context of the grsec build role, we'll default to
including the grsecurity patch version and timestamp, joined with ".",
to provide semver compatibility within the revision. This won't work
automatically for situations in which "10.00.Custom" is already active,
however, since 10 > 3.1 (the current grsecurity patch version). Thus the
override capability, so e.g. "11.00.Custom" would provide the necessary
semver upgrade path.