Bumps revision in SecureDrop kernel version
The previous kernel, currently in use by production SecureDrop
instances, shared the same kernel source version, 3.14.79. The default
revision for manual compiles supplied by `make-kpkg` is "10.00.Custom",
according to the make-kpkg man page, but the kernel version is prepended
automatically if you don't use the `--revision` flag, so it's actually
"3.14.79-grsec-10.00.Custom".

In order to ensure a smooth upgrade flow, we need to provide semver
compatibility with the previous version. Using "11.00.Custom" suffices.
Given the lack of automatic prepending when the `--revision` flag is
used, we must specify the entire kernel version as a prefix, i.e.
"3.14.79-grsec-11.00.Custom". Then we get sane version parsing according
to Debian package guidelines.

See the Debian Policy Guide for more info on how the version field is
parsed [0].

[0] https://www.debian.org/doc/debian-policy/ch-controlfields.html#s-f-Version