Enforces specific build targets for SecureDrop
Historically we've been maintaining both kernel image and header
packages, and posting them to the FPF apt repo (apt.freedom.press),
for use with SecureDrop. Only the image is technically necessary, but to
maintain backwards compatibility, let's ensure that the headers are
built as part of the SecureDrop-specific build playbook.

In addition, we must maintain the source packages, although at this time
we are not providing those publicly in the apt repository. Offers of source
for the hosted binaries will be available in the SecureDrop project
repo.