Updates SecureDrop kernel build docs
As often happens with technical documentation, the build guide for
SecureDrop kernels was out of date. Two major problems stood out:

  1. FPF staff with access to grsecurity portals couldn't follow
     the steps (recent PRs to this repo have adjusted the flow).
  2. SecureDrop Admins wishing to compile a kernel image from the
     source package would have a hard time reconstructing the steps.

Point 1) was rather easy to solve, given that the recent PRs vastly
simplified the flow. Docs are now accurate and should work for
subsequent kernel builds the next time we perform an update for
SecureDrop servers.

Point 2) is a bit trickier, because we're trying to provide a solid
documentation flow for a use case we haven't ever directly addressed
before: Admins wanting to recompile the kernel from source to verify
that it is correct (or simply to exercise their rights under the GPL).
Will need to ping some community members for review here to validate
that the docs are both clear and adequate.