Removes support for "test" patches
The grsecurity team has discontinued maintenance of the test patches,
meaning that the unauthenticated access to grsecurity patch sets is no
longer possible. Updated the grsecurity_urls module to remove all
mention of the test patches, and modified the associated documentation
to list only "stable" and "stable2" as the viable options.