Supports redirects in grsecurity patch fetching tasks
Unfortunately falling back to good ol' curl instead of using the
built-in `get_url` module, due the latter lacking reliable support for
redirects. That shortcoming really surprises me, so prove me wrong if
you can, but the 3xxs I tested with were throwing errors. Switching to
curl with the `-L` flag resolved the problems.

Using `curl` means we lose access to the file metadata parameters, such
as user, group, mode, etc., but since most of the build role runs
without superuser privileges, and we don't need or want the patches to
be executable, these are acceptable tradeoffs.