Overhauls README documentation
Recent PRs have revised the main logic flow, requiring substantial
updates to the documentation. Of particular note is the lack of "test"
patches due to discontinuation, and the recent hard requirement for
grsecurity login credentials in order to access patches (regardless of
patch type). Since the build documentation was separately updated,
the intro to the README now references it explicitly, because it may
be useful to SecureDrop Admins researching the build process.